requirements.txt
aiofiles==23.2.1
aiogram==3.6.0
aiohttp==3.9.5
aiomysql==0.2.0
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.6.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
behave==1.2.6
billiard==4.2.0
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==42.0.7
cssselect==1.2.0
Django==3.2
django-adminlte3==0.1.6
djangorestframework==3.15.1
drf-yasg==1.21.7
exceptiongroup==1.2.1
execnet==2.1.1
fasteners==0.19
filelock==3.14.0
flake8==7.0.0
frozenlist==1.4.1
gunicorn==22.0.0
h11==0.14.0
idna==3.7
inflection==0.5.1
iniconfig==2.0.0
isort==5.13.2
kombu==5.3.7
magic-filter==1.0.12
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
multidict==6.0.5
mysqlclient==2.2.4
outcome==1.3.0.post0
packaging==24.0
parameterized==0.9.0
parse==1.20.1
parse-type==0.6.2
pdbp==1.5.0
platformdirs==4.2.2
pluggy==1.5.0
prompt-toolkit==3.0.43
py==1.11.0
pycodestyle==2.11.1
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
pyflakes==3.2.0
Pygments==2.18.0
PyMySQL==1.1.0
pynose==1.5.1
pyotp==2.9.0
PySocks==1.7.1
pytest==8.2.0
pytest-html==2.0.1
pytest-metadata==3.1.1
pytest-ordering==0.6
pytest-rerunfailures==14.0
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.4
requests==2.31.0
rich==13.7.1
sbvirtualdisplay==1.3.0
selenium==4.21.0
seleniumbase==4.27.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.5.0
tabcompleter==1.3.0
tomli==2.0.1
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
wsproto==1.2.0
yarl==1.9.4
